Move pages to separate folder
